+++ src/app/organizer/organizer.ts
@@ -1,7 +1,0 @@
-export interface Organizer {
-  clientId?: string;
-  organizerId?: string;
-  name: string;
-  createdOn?: string;
-  lastUpdatedOn?: string;
-}